Refactoring names of files and folders for news and bonds
