+++ src/App.tsx
@@ -3,7 +3,7 @@ import Frame from './components/frame';
 import Commodity from './components/comodities/comodities';
 import CurrencyExchange from './components/currency/currency';
 import NewsComponent from './components/news/component';
-import BondsComponent from './components/bonds/component';
+import BondsComponent from './components/bonds/mainComponent';
 import { StockTable, GAINERS, LOSERS } from './components/stocktable';
 
 function App() {
