+++ src/bonds/bondsFunctionality.ts
