+++ src/news/newsFunctionality.ts
