+++ tsconfig.json
@@ -22,5 +22,5 @@
   },
   "include": [
     "src"
-, "src/tests/news.test.ts"  ]
+, "src/tests/newsFunctionality.test.ts"  ]
 }
